+++ src/infrastructure/databases/models/profiles.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, String, Integer, DateTime, Boolean
+from sqlalchemy import Column, String, BigInteger, DateTime, Boolean
 from src.infrastructure.databases import BaseModel
 
 
@@ -10,13 +10,12 @@ class Profile(BaseModel):
 
     __tablename__ = "profiles"
 
-    id: int = Column(Integer, primary_key=True, autoincrement=True)
+    id: int = Column(BigInteger, primary_key=True, autoincrement=True)
     phone: str = Column(String(11), nullable=True)
     address: str = Column(String(100), nullable=True)
     created_at: datetime = Column(DateTime, default=datetime.now, index=True)
     updated_at: datetime = Column(DateTime, nullable=True)
     is_deleted: bool = Column(Boolean, default=False, index=True)
-
 
     def __repr__(self) -> str:
         """
